Server side changes for API call completion.
Added changes to server side code to correctly route data requests to virtual database.
