+++ server/data/virtualDatabase.js
@@ -1,15 +1,11 @@
 const { MongoClient } = require('mongodb');
 const _ = import('lodash');
 
-// depreciate
-const uniqueCongress = require('./tempPoliticianData.js');
-
 // public functions
 module.exports = {
     // DESCRIPTION: main public function that gets sent a data request and returns a json variable
     // PROTOTYPE: sent a json w/ type of request and any data needed to make request
-    getData: async function(dataSlug) {        
-        console.log(dataSlug);
+    getData: async function (dataSlug) {
         const data = await getDataVirtualDatabase(dataSlug);
         return data;
     }
@@ -47,14 +43,20 @@ module.exports = {
 
 // DESCRIPTION: determines request type and routes appropriately, there should be no data processing in this function
 // PROTOTYPE: input dataSlug, return data object
-async function getDataVirtualDatabase(dataSlug){
+async function getDataVirtualDatabase(dataSlug) {
     // determine request type
 
     // sample/depreciated data request using hardcoded json
     if (dataSlug.requestType == "uniqueCongress") {
-        return uniqueCongress;
+        return getPoliticians();
+    } else if (dataSlug.requestType == "congressById") {
+        return getPolitician(dataSlug.requestData.id);
     } else if (dataSlug.requestType == "testMongo") {
         return await getSenatorByLast(dataSlug.requestData.congressLastName);
+    } else if (dataSlug.requestType == "testRequest") {
+        return {
+            test: "response1"
+        }
     }
 }
 
@@ -68,24 +70,24 @@ async function getDataVirtualDatabase(dataSlug){
 
 // DESCRIPTION: takes a mongodb call
 // PROTOTYPE:
-async function getMongoRequest(dataSlug){
+async function getMongoRequest(dataSlug) {
 }
 
 // DESCRIPTION: takes an API call to ProPublica API
 // PROTOTYPE:
-async function getProPublicaRequest(dataSlug){
+async function getProPublicaRequest(dataSlug) {
 }
 
 // DESCRIPTION: takes an API call to Polygon.io
 // PROTOTYPE:
-async function getPolygon(dataSlug){
+async function getPolygon(dataSlug) {
 }
 
 /*
     DEPRECIATED CALLS: phase these out, however good call structure for interim
 */
 
-async function getSenatorByLast(senator){
+async function getSenatorByLast(senator) {
     const uri = "mongodb://localhost:27017";
     const client = new MongoClient(uri);
 
@@ -103,7 +105,7 @@ async function getSenatorByLast(senator){
 }
 
 async function findOneListingByName(client, nameOfListing) {
-    const result = await client.db("member_sorted").collection("stockwatcher").findOne({last_name: nameOfListing});
+    const result = await client.db("member_sorted").collection("stockwatcher").findOne({ last_name: nameOfListing });
 
     if (result) {
         console.log(`Found a listing in the collection with the name '${nameOfListing}'`);
@@ -112,4 +114,40 @@ async function findOneListingByName(client, nameOfListing) {
     } else {
         console.log(`No listings found with the name of ${nameOfListing}`);
     }
+}
+
+let politicians = [
+    {
+        name: "Bernie Sanders",
+        state: "VT",
+        congressType: "Senator",
+        party: "D",
+        id: 1,
+    },
+    {
+        name: "Mitch McConnell",
+        state: "KY",
+        congressType: "Senator",
+        party: "R",
+        id: 2,
+    },
+    {
+        name: "Abraham Lincoln",
+        state: "DC",
+        congressType: "Representative",
+        party: "N",
+        id: 3,
+    },
+];
+
+// return array of json objects of all politiicans
+function getPoliticians() {
+    return politicians;
+}
+
+// return a specific politician id
+function getPolitician(id) {
+    return politicians.find(
+        (politician) => politician.id === id
+    );
 }
