+++ server/server.js
@@ -1,11 +1,13 @@
 const express = require('express');
+var cors = require('cors');
 const app = express();
 
 const virtualDatabase = require('./data/virtualDatabase.js');
-// depreciate
-const tempPoliticianData = require('./data/virtualDatabase.js');
 
 // EXPRESS CODE
+// get cors working
+app.use(cors());
+// set up listening on port
 app.listen(5000, () => console.log('listening at 5000'));
 app.use(express.static('public'));
 
@@ -14,25 +16,15 @@ app.use(express.json({ limit: '1mb' }));
 
 // API routing to virtual database
 app.post('/api', async (request, response) => {
-    // send data request to virtual database
+    console.log('I got a request to test!');
+    console.log(request.body);
+    // send request to virtual database, and get constructed object back
     const data = await virtualDatabase.getData(request.body);
     // return data
     response.json({
         status: 'success',
         message: data,
     });
-});
-
-// depreciate
-// API routing test response, just used to make sure things are hooked up
-app.post('/apiTEST', async (request, response) => {
-  console.log('I got a request to test!');
-  console.log(request.body);
-  // get sample data from tempPoliticianData
-  const data = getPoliticians();
-  // return data
-  response.json({
-      status: 'success',
-      message: data,
-  });
+    console.log("Returned data object!");
+    console.log(data);
 });
